requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Pillow==9.3.0
pytz==2022.2.1
razorpay==1.3.0
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
